commit spring-cloud_microservices shopping-cart fix v8 v11

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/api/shoppingCart/ShoppingCartApi.java
@@ -22,7 +22,7 @@ public interface ShoppingCartApi {
 
     @PutMapping
     ShoppingCartDto addProductToShoppingCart(
-            @RequestBody @NotEmpty Map<@NotNull UUID, @NotNull @Positive Long> products,
+            @RequestBody @NotEmpty Map<@NotNull UUID, @NotNull @Positive Integer> products,
             @RequestParam @NotBlank String username);
 
     @DeleteMapping
